pyproject.toml
[project]
name = "leads-40"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-search-results>=2.4.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5"
]
